+++ changelog.md
@@ -152,3 +152,5 @@
 # 0.5.1
 
 * Added constructor `ModbusTcpServer(ModbusInterface *device)` for compatibilty with prev versions
+* Added `ModbusClientPort::frameRequest` function which makes raw request within current frame
+  excluding unit, function, TCP-prefix, CRC/LRC checksums.